# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Greater Bayelsa leadership development platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List public posts
      parameters:
        - name: category
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post (HQ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post (HQ)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (HQ)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------- Community updates ----------
  /updates:
    get:
      operationId: listUpdates
      tags: [updates]
      summary: Community updates feed within scope, urgent first
      responses:
        "200":
          description: Updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityUpdate"
    post:
      operationId: createUpdate
      tags: [updates]
      summary: Post a community update (leaders and HQ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityUpdateInput"
      responses:
        "201":
          description: Created update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityUpdate"

  # ---------- Notifications / messaging ----------
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark own notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages:
    post:
      operationId: sendMessage
      tags: [notifications]
      summary: Send bulk or individual message to members (HQ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageSendInput"
      responses:
        "201":
          description: Message fan-out result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResult"

  # ---------- Feedback ----------
  /feedback:
    get:
      operationId: listFeedback
      tags: [feedback]
      summary: List member feedback reports (HQ) or own reports (member)
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedbackReport"
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit a direct report to headquarters (member)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackReport"
  /feedback/{id}:
    patch:
      operationId: updateFeedback
      tags: [feedback]
      summary: Respond to / resolve a feedback report (HQ)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------- Storage ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Post:
      type: object
      required: [id, title, body, category, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        body: { type: string }
        category: { type: string }
        imageUrl: { type: [string, "null"] }
        createdAt: { type: string }
    PostInput:
      type: object
      required: [title, body, category]
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
        category: { type: string, enum: [news, impact, development] }
        imageUrl: { type: [string, "null"] }
    PostUpdate:
      type: object
      properties:
        title: { type: string }
        body: { type: string }
        category: { type: string, enum: [news, impact, development] }
        imageUrl: { type: [string, "null"] }
    CommunityUpdate:
      type: object
      required: [id, title, body, urgent, createdAt]
      properties:
        id: { type: integer }
        villageId: { type: [integer, "null"] }
        villageName: { type: [string, "null"] }
        authorId: { type: integer }
        authorName: { type: [string, "null"] }
        title: { type: string }
        body: { type: string }
        urgent: { type: boolean }
        createdAt: { type: string }
    CommunityUpdateInput:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
        urgent: { type: boolean }
        villageId: { type: [integer, "null"] }
    Notification:
      type: object
      required: [id, title, body, kind, read, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        body: { type: string }
        kind: { type: string }
        read: { type: boolean }
        createdAt: { type: string }
    MessageSendInput:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
        userIds:
          type: [array, "null"]
          items: { type: integer }
        villageId: { type: [integer, "null"] }
    SendMessageResult:
      type: object
      required: [notified]
      properties:
        notified: { type: integer }
    FeedbackReport:
      type: object
      required: [id, userId, category, body, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        memberName: { type: [string, "null"] }
        membershipCode: { type: [string, "null"] }
        villageName: { type: [string, "null"] }
        category: { type: string }
        body: { type: string }
        status: { type: string }
        response: { type: [string, "null"] }
        createdAt: { type: string }
    FeedbackInput:
      type: object
      required: [category, body]
      properties:
        category:
          type: string
          enum: [leadership_concern, security_issue, dues_dispute, other]
        body: { type: string, minLength: 1 }
    FeedbackUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [open, reviewed, resolved] }
        response: { type: [string, "null"] }
...